requirements.txt
python-telegram-bot==20.0
pycoingecko==3.0.0
pytest-asyncio
sys
os
